requirements.txt
celery==5.6.2
redis==5.2.1
requests==2.32.5
python-dotenv==1.0.1
flower==2.0.1
python-multipart>=0.0.22
protobuf>=5.29.3
fastapi==0.115.0
uvicorn==0.31.0
python-telegram-bot==21.6
py-localtunnel==1.0.3
